Renamed package name and util class name
